move auth into users.py
